Smart_CCTV_Project/requirements.txt
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62  # If you're running on a server without GUI
scikit-image==0.19.3
numpy==1.21.2
Pillow==8.4.0
beepy==1.0.1
